requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
PyJWT==2.7.0
pytesseract==0.3.10
python-docx==0.8.11
requests==2.31.0
SQLAlchemy==2.0.17
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
